requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-admin==2.0.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-excel-response2==3.0.6
django-six==1.0.5
django-timezone-field==6.1.0
excel-base==1.0.4
facebook-sdk==3.1.0
idna==3.6
isoweek==1.3.3
kombu==5.3.5
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
redis==5.0.1
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlwt==1.3.0
